requirements/requirements.txt
Django==4.2.7
psycopg>=3.1.8
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
django-filter
